chore(plugins): rename files to kebab-case across remaining plugins
