+++ packages/googlemeet/endpoints/smart-notes.ts
