+++ packages/insightoai/endpoints/index.ts
@@ -5,7 +5,7 @@ import * as Datasources from './datasources';
 import * as Forms from './forms';
 import * as Providers from './providers';
 import * as Tools from './tools';
-import * as WebhooksTelephony from './webhooksTelephony';
+import * as WebhooksTelephony from './webhooks-telephony';
 import * as Widgets from './widgets';
 
 export {
